qwen4exp: full forward graph and autoregressive generate
Ports the complete Qwen3.8-Flash-Next trunk from upstream qwen4exp.cpp:

- build_linear_attn  gated delta net (36 layers): fused qkv + z gate,
  beta/alpha/dt_bias/ssm_a gates, depthwise causal conv over the prepended
  history (ggml_ssm_conv), L2-normed q/k, ggml_gated_delta_net (which gets
  the tiled RDNA3.5 kernel), sigmoid-gated RMSNorm, ssm_out.
- build_full_attn  dense GQA (12 layers): interleaved q|gate split,
  q/k RMSNorm, partial I-MRoPE (GGML_ROPE_TYPE_MROPE, sections 11/11/10/0),
  F16 KV write into the persistent cache, ggml_flash_attn_ext, sigmoid gate.
- build_ple  per-layer n-gram embedding on the PLE layer: host-side hash
  (u64 layer multipliers, per-head vocab/offset) gathered through
  Qwen4ExpPleReader, key/query grouped RMSNorm, signed-sqrt sigmoid gate,
  dilated depthwise conv with the persistent history.
- qwen4exp_forward  embed -> hc_init -> per layer {PLE, HC-attn mix,
  attention, HC combine, HC-ffn mix, MoE, HC combine} -> output HC mix ->
  lm_head; chunked causal mask; gallocr; logits read-back.
- loader: PLE metadata arrays are u64 (head_offsets, head_vocab_sizes,
  layer_multipliers); read them as such. This was a load-time rejection.
- backend generate_impl: prefill in chunks then greedy/sampled AR decode
  through the shared sample_logits chain, resetting the recurrent state and
  PLE window per request.

Compiles clean. Numerics are unverified until this runs on the Halo box.

+++ server/src/qwen4exp/qwen4exp_internal.h
@@ -205,9 +205,9 @@ struct Qwen4ExpWeights {
     int ple_head_dim          = 0;      // = embedding_length_per_layer_input
     int ple_n_heads           = 0;      // = (ngram_size - 1) * heads_per_ngram
     std::vector<int32_t> ple_layer_ids;
-    std::vector<int32_t> ple_head_offsets;
-    std::vector<int32_t> ple_head_vocab_sizes;
-    std::vector<float>   ple_layer_multipliers;
+    std::vector<int64_t> ple_head_offsets;       // u64 in GGUF
+    std::vector<int64_t> ple_head_vocab_sizes;   // u64 in GGUF
+    std::vector<uint64_t> ple_layer_multipliers; // u64 in GGUF
     int32_t ple_eos_token_id  = -1;
     int32_t ple_image_token_id = -1;
